% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPfunction.R
\name{calcBt}
\alias{calcBt}
\title{Calculate biomass in next year using estimated carrying capacity and growth rate}
\usage{
calcBt(Bt, Ct, r, K)
}
\arguments{
\item{Bt}{scalar. Current level of biomass from which to calculate next year biomass}

\item{Ct}{scalar. Current level of catches from which to calculate next year biomass}

\item{r}{scalar. Growth rate}

\item{K}{scalar. Carrying capacity}
}
\value{
scalar. Biomass in next year
}
\description{
Calculate biomass in next year using estimated carrying capacity and growth rate
}
\references{
Copied from Rajeev and Divya's RMSE package
}
