% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{Flinear}
\alias{Flinear}
\title{Title This function calculates F for a linear HCR}
\usage{
Flinear(Bt, Blim, Btrigger, Ftarget, Fmin = 0.001, ...)
}
\arguments{
\item{Bt}{: stock biomass at time t}

\item{Blim}{: Biomass limit reference point (upper boundary for critical zone)}

\item{Btrigger}{: Biomass trigger reference point (upper boundary for cautious zone)}

\item{Ftarget}{: Target Fishing mortality}

\item{Fmin}{: defaut Fmin=0.001.  A very small value of F set at the critical zone. Setting Fmin=0, sometime causes trouble}
}
\description{
Title This function calculates F for a linear HCR
}
\references{
copied from Kumar and Varkey RMSE package
}
