% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{fnCall}
\alias{fnCall}
\title{Function caller.}
\usage{
fnCall(fnName, ...)
}
\arguments{
\item{fnName}{Name of the function to be called in quote}

\item{...}{arguments require for the function to be called}
}
\value{
Output of the actual function
}
\description{
Call any R function by the function name in quote followed by that function args, separated by commas
}
\examples{
fnName="mean"
x=c(2,5,6,NA)
match.fun(fnName)(x,na.rm=TRUE)

}
\references{
copied from Kumar and Varkey RMSE package
}
