% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{Fsigmoid}
\alias{Fsigmoid}
\title{Title This function calculates F for a sigmoid HCR}
\usage{
Fsigmoid(Bt, Blim, Btrigger, x50, Ftarget, k = 1, Fmin = 0.001, ...)
}
\arguments{
\item{Bt}{: stock biomass at time t}

\item{Blim}{: Biomass limit reference point (upper boundary for critical zone)}

\item{Btrigger}{: Biomass trigger reference point (upper boundary for cautious zone)}

\item{x50}{: Distance from Blim where F=0.5*Ftarget expressed as fraction of cautious zone biomass range. For Upper leaf: 0>x50>0.5 and for Lower leaf: 0.5>x50>1}

\item{Ftarget}{: Target Fishing mortality}

\item{k}{: shape parameter for the sigmoid HCR}

\item{Fmin}{: defaut Fmin=0.001.  A very small value of F set at the critical zone. Setting Fmin=0, sometime causes trouble}

\item{...}{:useful if uses as a nested function (inside a function)}
}
\description{
Title This function calculates F for a sigmoid HCR
}
\references{
copied from Kumar and Varkey RMSE package
}
