% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPfunction.R
\name{getOutputs}
\alias{getOutputs}
\title{Gets output from Jags object}
\usage{
getOutputs(modelRunOutput)
}
\arguments{
\item{modelRunOutput}{JAGS model output object}
}
\value{
summarized model output
}
\description{
Gets output from Jags object
}
