% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{plotLeaf}
\alias{plotLeaf}
\title{Plot PA leaf}
\usage{
plotLeaf(
  Blim = 12240,
  Btrigger = 22949,
  Ftarget = 0.269,
  x50UL = 0.25,
  x50ML = 0.5,
  x50LL = 0.75,
  Fmin = 0.001
)
}
\arguments{
\item{Blim}{B}

\item{Btrigger}{: Biomass trigger reference point (upper boundary for cautious zone)}

\item{Ftarget}{: Target Fishing mortality}

\item{x50UL}{Distance from Blim where F=0.5*Ftarget expressed as fraction of cautious zone biomass range. For Upper leaf: 0>x50>0.5}

\item{x50ML}{Distance from Blim where F=0.5*Ftarget expressed as fraction of cautious zone biomass range. For mid leaf: x50=0.5}

\item{x50LL}{Distance from Blim where F=0.5*Ftarget expressed as fraction of cautious zone biomass range. Lower leaf: 0.5>x50>1}

\item{Fmin}{: defaut Fmin=0.001.  A very small value of F set at the critical zone. Setting Fmin=0, sometime causes trouble}
}
\description{
Plot PA leaf
}
\references{
copied from Kumar and Varkey RMSE package
}
