% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{plotLeafStock}
\alias{plotLeafStock}
\title{Title}
\usage{
plotLeafStock(
  Blim = 12240,
  Btrigger = 22949,
  Ftarget = 0.269,
  x50UL = 0.25,
  x50ML = 0.5,
  x50LL = 0.75,
  Fmin = 0.001,
  stockval = NULL,
  stockF = NULL,
  xtxt = "SSB"
)
}
\arguments{
\item{stockval}{Current stock level to be indicated}
}
\description{
Title
}
\references{
copied from Kumar and Varkey RMSE package
}
