% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{Runs Jags model}
\usage{
runModel(
  modeltxt,
  data,
  inits,
  parameters,
  n.chains,
  n.adapt,
  n.burnin,
  n.iter,
  n.thin
)
}
\arguments{
\item{n.thin}{}
}
\description{
Runs Jags model
}
